services/worker/requirements.txt
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
ultralytics==8.0.110
Pillow==10.0.0
numpy==1.24.3
onnxruntime==1.17.3
